legacy/requirements.txt
# Core dependencies for Mental Health Chatbot
streamlit>=1.28.0
openai>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0

# Agent optimization and training
agentlightning
# parlant>=1.0.0  # DISABLED: Dependency conflict with Starlette/FastAPI
# starlette<0.37.0  # Required for parlant compatibility
